Refactor to make script that first runs backfill pipeline and then live
